implement barebones sync node command
This is enough to rsync the necessary files to a node and run chef-solo.
It requires chef-solo 0.10, but doesn't check for that.

Because this is a PITA to test, there is no automated testing.
Eventually, I should mock Net::OpenSSH and pretend things work
or fail, but that's not really a great test either.